Update project for GitHub Pages deployment
This commit aims to prepare the project for deployment on GitHub Pages. Global CSS settings were removed and the Next.js configuration was updated accordingly, including recent changes to output and basePath. The README file was simplified and now only provides project identification and a link to the live site. A build step was also added to the package.json, and an AppVeyor configuration file was introduced for continuous integration. Last but not least, .idea was added to .gitignore to ensure that IDE configs are not version control tracked.